Add new task and update task data
This commit adds a new task to the taskData array in the task.data.ts file. The new task has an ID of 3, a title of 'Tarea C', a limit date of '2024-09-31', is not completed, has a priority of 3, and includes two persons with their respective names, ages, and skills.

Additionally, a new priorityName pipe is added in the priority-name.pipe.ts file. This pipe transforms a priority number into a corresponding priority name, such as 'Baja' for 1, 'Media' for 2, 'Alta' for 3, and 'Desconocida' for any other value.

These changes enhance the functionality of the application by allowing the addition of new tasks and providing a way to display the priority name of each task.